zerver: Add command to sync groups with streams.
This command adds user to streams based on a configuration file.
It can be used like:

./manage.py apply_group_stream_mapping -r 2 -c /home/zulip/mapping.json

The format of `channelmapping.json` is the same as in Rocketchat:

```json
{
  "group1" : [ "stream-1", "another stream" ],
  "another group" : [ "stream2", "another stream"],
}
```

The command does not automatically create streams. It also only
synchronizes private streams. Using the `--remove-users` flag also
removes all user that are not inside the required groups.

The logging output without `--remove-users` flag also shows users
that should be removed. This can be used for testing purposes.